restructures Hero component
The Hero component has been restructured with a more meaningful collocation approach.
