+++ src/components/Hero/subComponents/Hero.js
@@ -1,11 +1,14 @@
 import React from 'react';
+
 // hero section background image 
-import background from '.././../resources/images/blackboardBgOptimized.jpg'; 
+//TODO optimize these images - what's the best route to take? Research 🤓🤓
+import background from './blackboardBgOptimized.jpg'; 
+
 //* material ui components 
 import { Box, Typography } from '@material-ui/core';
 // * material ui hooks
 import { withStyles } from '@material-ui/core/styles'; 
-import styles from '../../styles/LandingStyles'; 
+import styles from './HeroStyles'; 
 
 function Hero({ classes }) {
 
@@ -25,7 +28,7 @@ export default withStyles(styles)(Hero);
 // There was too much element nesting going on before, and an overly complicated approach to structuring such a simplistic part of the page 
 // ! Wireframes can be seen at: https://trello.com/c/1ky1rTXR 
 
-// New flow: 
+// Flow: 
 // * 1 parent element - the MUI Box component, a generic and cutomizable div
 // * Child elements: 2 Typography components, 1 Button component 
 //! 3.26 - I removed the register button from this little hero section. I don't know where to put it right this minute, and I still need to build a proper register form. 😅
