+++ src/components/Hero/subComponents/HeroDetailsStyles.js
