+++ src/components/Hero/subComponents/HeroStyles.js
