+++ src/components/Hero/subComponents/blackboardBgOptimized.jpg
