+++ src/components/Hero/subComponents/forgetting-curve-graphic.jpg
