+++ src/components/Hero/subComponents/spaced-rep-graphic.jpg
